# integrations/chatgpt/openapi.yaml
openapi: 3.1.0
info:
  title: SovereignAI
  description: >
    Talk to the user's private, self-hosted sovereign AI: ask it questions
    (answered with the user's personas, long-term memory, and private knowledge
    base), search that knowledge base, and store memories or documents.
  version: 0.1.0
servers:
  - url: https://YOUR-PUBLIC-TUNNEL-URL
    description: Public tunnel to your SovereignAI server (see README)
paths:
  /api/ask:
    post:
      operationId: askSovereign
      summary: Ask the user's sovereign AI a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  description: The question or request
                persona:
                  type: string
                  description: Optional persona name (e.g. Assistant, Engineer, Archivist)
      responses:
        '200':
          description: The AI's answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  persona: { type: string }
                  model: { type: string }
  /api/search:
    get:
      operationId: searchKnowledge
      summary: Search the user's private knowledge base
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Relevant excerpts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    document: { type: string }
                    content: { type: string }
                    score: { type: number }
  /api/memories:
    post:
      operationId: addMemory
      summary: Store a long-term memory note in the user's sovereign AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        '200':
          description: Stored memory
  /api/documents:
    post:
      operationId: addKnowledge
      summary: Save a document into the user's private knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string }
                content: { type: string }
      responses:
        '200':
          description: Ingested document
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []